refactor: 重构抢券预约提醒数据库表结构，方便后续扩展

+++ engine/src/main/java/com/mall/cqupt/engine/dto/req/CouponTemplateRemindCreateReqDTO.java
@@ -26,7 +26,7 @@ public class CouponTemplateRemindCreateReqDTO {
      * 预约抢券时间点，可以接受开抢前五分钟到前一小时的预约，五分钟一个维度，以位图的形式，比如预约前十五分钟，就是1 << ((15 / 5) - 1)，也就是4(二进制100)
      */
     @Schema(description = "预约时间点", example = "4", required = true)
-    private Integer appointmentBitMap;
+    private Long appointmentBitMap;
 
     /**
      * 提醒方式
